requirements.txt
fastapi
uvicorn[standard]
pillow
pytesseract
pdf2image
opencv-python-headless
python-multipart
numpy
pandas
regex
python-Levenshtein
fuzzywuzzy
requests
